requirements.txt
numpy
torch
torchvision
scikit-learn
joblib
Pillow
opencv-python
mediapipe
